refactor dir.
