+++ lib/screens/boards/board.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import '../../models/project.dart';
-import 'notice.dart';
-import 'tasks.dart';
-import 'threads.dart';
+import 'notice/notice.dart';
+import 'tasks/tasks.dart';
+import 'threads/threads.dart';
 import 'package:provider/provider.dart';
 import '../../providers/boards/boards.dart';
 import 'board_title/board_title.dart';
