+++ lib/screens/boards/notice/notice.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import '../../commons/circular_border_container.dart';
-import '../../models/boards/thread.dart' as ThreadModel;
-import '../../commons/profile_thumbnail.dart';
+import '../../../commons/circular_border_container.dart';
+import '../../../models/boards/thread.dart' as ThreadModel;
+import '../../../commons/profile_thumbnail.dart';
 import 'package:provider/provider.dart';
-import '../../providers/boards/boards.dart';
+import '../../../providers/boards/boards.dart';
 
 class Notice extends StatelessWidget {
   final ThreadModel.Thread notice;
